% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishersTestExceedance.R
\name{fishersTestExceedance}
\alias{fishersTestExceedance}
\title{Fishers test to compare two mean squared error rates.}
\usage{
fishersTestExceedance(Y, predicted, E, n, p, alpha, nSeqTests, verbose)
}
\arguments{
\item{verbose}{}
}
\value{
A list with the p-value for the test.
}
\description{
Tests whether ...
}
