% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KCI.R
\name{KCI}
\alias{KCI}
\title{Kernel conditional independence test.}
\usage{
KCI(Y, E, X, width = 0, alpha = 0.05, unbiased = FALSE, approx = TRUE,
  bootstrap = TRUE, isEcategorical = FALSE, nRepBs = 500,
  lambda = 0.001, thresh = 1e-05, numEig = length(Y), verbose = FALSE)
}
\arguments{
\item{width}{Kernel width}

\item{alpha}{Significance level}

\item{verbose}{}
}
\value{
A list with the following entries:
\itemize{
 \item \code{testStatistic} the statistic Tr(K_{\ddot{X}|Z} * K_{Y|Z})
 \item \code{criticalValue} the critical point at the p-value equal to alpha obtained by bootstrapping.
 \item \code{pValue} the p value obtained by bootstrapping.
 \item \code{criticalValueApprox} the critical value obtained by Gamma approximation.
 \item \code{pValueApprox} the p-value obtained by Gamma approximation.
 }
}
\description{
Tests the null hypothesis that Y and E are independent given X.
}
\examples{
n <- 100
E <- rnorm(n)
X <- 4 + 2 * E + rnorm(n)
Y <- 3 * (X)^2 + rnorm(n)
KCI(Y, E, X)
KCI(Y, X, E)

}
