% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ksResidualDistributions.R
\name{ksResidualDistributions}
\alias{ksResidualDistributions}
\title{KS test to compare residual distributions}
\usage{
ksResidualDistributions(Y, predicted, E, n, p, alpha, nSeqTests, verbose)
}
\arguments{
\item{verbose}{}
}
\value{
A list with the p-value for the test.
}
\description{
Tests whether ...
}
