% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leveneAndWilcoxResidualDistributions.R
\name{leveneAndWilcoxResidualDistributions}
\alias{leveneAndWilcoxResidualDistributions}
\title{test to compare residual distributions}
\usage{
leveneAndWilcoxResidualDistributions(Y, predicted, E, n, p, alpha, nSeqTests,
  verbose)
}
\arguments{
\item{verbose}{}
}
\value{
A list with the p-value for the test.
}
\description{
Tests whether ...
}
