% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propTestTargetE.R
\name{propTestTargetE}
\alias{propTestTargetE}
\title{test to compare two misclassification rates.}
\usage{
propTestTargetE(E, predictedOnlyX, predictedXY, n, p, alpha, nSeqTests, verbose)
}
\arguments{
\item{verbose}{}
}
\value{
A list with the p-value for the test.
}
\description{
Tests whether ...
}
