% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wilcoxTestTargetY.R
\name{wilcoxTestTargetY}
\alias{wilcoxTestTargetY}
\title{Wilcoxon test to compare two mean squared error rates.}
\usage{
wilcoxTestTargetY(Y, predictedOnlyX, predictedXE, n, p, alpha, nSeqTests,
  verbose, ...)
}
\arguments{
\item{verbose}{}
}
\value{
A list with the p-value for the test.
}
\description{
Tests whether ...
}
