% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishersTestExceedance.R
\name{fishersTestExceedance}
\alias{fishersTestExceedance}
\title{Fishers test to test whether the exceedance of the conditional quantiles
is independent of the categorical variable E.}
\usage{
fishersTestExceedance(Y, predicted, E, adjFactor, verbose)
}
\arguments{
\item{Y}{An n-dimensional vector.}

\item{predicted}{A matrix with n rows. The columns contain predictions for different
conditional quantiles of Y|X.}

\item{E}{An n-dimensional vector, defining the grouping.}

\item{adjFactor}{Bonferroni adjustment factor for p-value if multiple tests were performed.}

\item{verbose}{Set to \code{TRUE} if output should be printed.}
}
\value{
A list with the p-value for the test.
}
\description{
Used as a subroutine in \code{InvariantConditionalQuantilePrediction}
to test whether the exceedance of the conditional quantiles
is independent of the categorical variable E.
}
