% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fTestTargetY.R
\name{fTestTargetY}
\alias{fTestTargetY}
\title{F test to compare two fits...}
\usage{
fTestTargetY(Y, predictedOnlyX, predictedXE, n, p, alpha, nSeqTests, verbose,
  ...)
}
\arguments{
\item{verbose}{}
}
\value{
A list with the p-value for the test.
}
\description{
Tests whether ...
}
